Add new git_worktree_name and git_worktree_path functions
The tests for git_worktree_path do string comparisons on two paths and on macOS the default temp/var dirs are usually symlinked. This made direct string comparisons impossible as the path as stored in the worktree structure had links already resolved. So the sandbox.h changes attempt to get the realpath for the tmp directory on macOS.