mem: create new interface to handle memory mappings
Throughout our codebase, there are quite a lot of instances where we use
the `git_map` structure to track arbitrary data. Originally, this
interface is only an abstraction such that we can emulate the POSIX
`mmap` interface across different platforms, most importantly Win32.
As such, this interface is a poor fit to keep track of arbitrary memory.

Instead, we now introduce a new structure `git_mem`. Basically, this
structure holds a pair of data and length members to allow a user to
read and/or write arbitrary memory. In contrast to the old interface,
though, `git_mem` will keep track of the way it has been initialized and
then correctly handle disposure of that memory. There are currently
three major modes through which to acquire the memory mapping:

- `git_mem_from_fd` and `git_mem_from_path` will create a read-only
  mapping of a given file into memory. It is thus a high-level `mmap`
  interface.

- `git_mem_from_data` will take a void pointer and length of the mapping
  and own it from thereon. Disposal will cause the pointer to be free'd.

- `git_mem_from_unowned` will take a void pointer and a length, too. In
  contrast to the previous function, though, the pointer is considered
  unowned and will not be free'd on disposal.