diff: use the new `git_mem` interface
The structure `git_diff_file_content` makes use of the `git_map`
structure to keep track of the file contents and the way it has been
loaded. Instead, use the new `git_mem` interface, which has been
designed for this exact purpose. This allows us to get rid of all the
accounting with regards to how we have allocated the `git_map` data.