indexer: use the new `git_mem` interface
When reading a pack header from an opened pack file, we use a temporary
`git_map`. While this is a "real" usecase for using a mmap, we are about
to remove the `git_map` interface in favor of using `git_mem` instead.
Switch the indexer over to use the new interface instead.