cmake: increase WIN32_WINNT to Vista
Increase the WIN32_WINNT level to 0x0600, which enables support for new
APIs from Windows 6.0 (Vista).  We had previously set this to 0x0501,
which was Windows XP.  Although we removed XP support many years ago,
there was no need to update this level previously.  We're doing so now
explicitly so that we can get support for the `CreateSymbolicLink` API.