examples: add example that resembles git-submodule(1)
Add an example that mimics parts of the interface of git-submodule(1).
Implemented are the subcommands "add", "init", "summary", "sync" and
"update".

A shortcoming of these examples is that for "init", we will always
re-initialize the submodule even if it already was initialized. This
results from the fact that there is no easy way to check if a submodule
has already been initialized (that is, a key "submodule.$SM.url" exists
in the ".git/config" file of the superproject). While there is the
`git_submodule_status` API, the only suspect of these status flags would
be `GIT_SUBMODULE_STATUS_IN_CONFIG`. But this flag in fact indicates
that a submodule has been initialized in the ".gitmodules" file, not in
".git/config". So users of libgit2 are kind of screwed in case they do
not want to write their own code querying ".git/config" for internal
variables.