utils: make sure memchr() is not invoked with NULL
Passing a null pointer as first argument to `memchr()` results in
undefined behavior.

In practice, this should not be a problem since, within libgit2,
`git__linenlen()` will be either called with a non-null `buffer` or with
both `buffer` and `buffer_len` being null. Any sane implementation of
`memchr()` should return `NULL` in this case without causing any
malicious or unintended effects.

However, it is still an UB. It will be reported by both dynamic and
static UB checkers. Hence, we add this additional check. And while we're
at it, we also introduce an assertion to catch calls with a null
`buffer` and a non-null `buffer_len`.